StudyController is fixed.

+++ src/main/java/org/example/dimsproject/Controller/StudyController.java
@@ -5,33 +5,27 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
+import javafx.stage.Popup;
 import javafx.stage.Stage;
+import javafx.stage.Window;
 import org.example.dimsproject.HelloApplication;
 import org.example.dimsproject.model.Study;
 import org.example.dimsproject.repository.StudyRepository;
+import org.example.dimsproject.utils.PopUp;
 
 import java.io.IOException;
+import java.util.Optional;
+
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
+import org.example.dimsproject.utils.PopUp;
 
 public class StudyController {
 
-    public StudyController(){
-        StudyRepository studyRepository = new StudyRepository();
-    }
 
     @FXML
     private Button savebtn;
-    @FXML
-    private Button updatebtn;
-    @FXML
-    private Button deletebtn;
-    @FXML
-    private Button clearbtn;
-    @FXML
-    private Button advisorpage;
-    @FXML
-    private Button closebtn;
-    @FXML
-    private Button fetchbtn;
+
     @FXML
     private TextField T1;
     @FXML
@@ -39,6 +33,10 @@ public class StudyController {
     @FXML
     private TextField T3;
 
+    private final StudyRepository studyRepository;
+    public StudyController(){
+         this.studyRepository = new StudyRepository();
+    }
     @FXML
     void close(ActionEvent event) {
         Stage stage = (Stage) savebtn.getScene().getWindow();
@@ -52,14 +50,42 @@ public class StudyController {
 
     }
     @FXML
-    void delete(ActionEvent event)
+    public void delete(ActionEvent event)
     {
-        StudyRepository studyRepository = new StudyRepository();
+        if (T1.getText().isBlank()){
+            PopUp.showPopup("Warning!","Id is mandatory!", Alert.AlertType.ERROR);
+            return;
+        }
         studyRepository.deleteById(Integer.parseInt(T1.getText()));
+        PopUp.showPopup("Success!","Study is deleted successfully!: "+T1.getText(), Alert.AlertType.WARNING);
+
+    }
+
+    private Study getStudy()
+    {
+        Study study;
+        if (T1.getText().isBlank() || T2.getText().isBlank() || T3.getText().isBlank())
+        {
+            PopUp.showPopup("Fields!","All fields mandatory!", Alert.AlertType.WARNING);
+            return null;
+        }
+        study = new Study();
+        study.setId(Integer.parseInt(T1.getText()));
+        study.setTitle(T2.getText());
+        study.setDescription(T3.getText());
+        return study;
     }
 
     @FXML
     void getAdvisorView(ActionEvent event) throws IOException {
+        Stage existingStage = getExistingAdvisorStage();
+
+        if (existingStage != null) {
+            existingStage.show();
+            existingStage.toFront();
+            return;
+        }
+
         Stage stage = new Stage();
         FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("advisor.fxml"));
         Scene scene = new Scene(fxmlLoader.load(), 550, 400);
@@ -68,39 +94,53 @@ public class StudyController {
         stage.show();
     }
     @FXML
+    private Stage getExistingAdvisorStage() {
+        for (Window window : Window.getWindows()) {
+            if (window instanceof Stage) {
+                Stage stage = (Stage) window;
+                if ("Advisor".equals(stage.getTitle())) {
+                    return stage;
+                }
+            }
+        }
+        return null;
+    }
+
+
+    @FXML
     void fetch(ActionEvent event) throws Exception {
-        StudyRepository studyRepository = new StudyRepository();
-        Study s = studyRepository.getStudyById(Integer.parseInt(T1.getText()));
-        if (s.getId()==0){
-            Alert alert = new Alert(Alert.AlertType.INFORMATION, "No such record!", ButtonType.CLOSE);
+        if (T1.getText().isBlank()){
+            PopUp.showPopup("Fields!", "All fields mandatory!", Alert.AlertType.ERROR);
+            return;
+        }
+        Optional<Study> c = Optional.ofNullable(studyRepository.getStudyById(Integer.parseInt(T1.getText())));
+        if (c.isEmpty()){
+            Alert alert = new Alert(Alert.AlertType.INFORMATION, "No such record!",ButtonType.CLOSE);
             alert.showAndWait();
             T1.setText("");
             T1.requestFocus();
         }
-        T2.setText(s.getTitle());
-        T3.setText(s.getDescription());
+        else {
+            T2.setText(c.get().getTitle());
+            T3.setText(c.get().getDescription());
+        }
 
     }
     @FXML
     void save(ActionEvent event) throws Exception
     {
-        StudyRepository studyRepository = new StudyRepository();
-        Study s = new Study();
-        s.setId(Integer.parseInt(T1.getText()));
-        s.setTitle(T2.getText());
-        s.setDescription(T3.getText());
-        studyRepository.saveNewStudy(s);
-
+        Study study = getStudy();
+        if (study == null) return;
+        studyRepository.saveNewStudy(study);
+        PopUp.showPopup("Success!","Study is created! :"+study.getId(), Alert.AlertType.INFORMATION);
     }
     @FXML
     void update(ActionEvent event) throws Exception
     {
-        StudyRepository studyRepository = new StudyRepository();
-        Study s = new Study();
-        s.setId(Integer.parseInt(T1.getText()));
-        s.setTitle(T2.getText());
-        s.setDescription(T3.getText());
-        studyRepository.updateStudy(s);
+        Study study = getStudy();
+        if (study == null) return;
+        studyRepository.updateStudy(study);
+        PopUp.showPopup("Success!","Study is updated successfully! :"+study.getId(), Alert.AlertType.INFORMATION);
 
     }
